Renamed modals file and migrated close btn back
