+++ src/components/pages/address-overview.tsx
@@ -5,7 +5,7 @@ import { useIntl } from "react-intl";
 import styled from "styled-components";
 import useSWR, { mutate } from "swr";
 
-import { Modal, ModalVariant } from "../modal";
+import { Modal, ModalVariant } from "../modals";
 import { Address } from "@src/@types/profile";
 import { PrimaryBadge } from "@src/components/badges";
 import { Box } from "@src/components/box";
